// frontend/src/client/types.gen.ts
export type ContactPublic = {
    first_name: string;
    last_name: string;
    title?: (string | null);
    email?: (string | null);
    phone?: (string | null);
    linkedin_url?: (string | null);
    notes?: (string | null);
    id: string;
    company_id: string;
    tenant_id: string;
    created_at?: (string | null);
    updated_at?: (string | null);
};